(PC-2448) code review
