+++ src/components/pages/search/FilterControls/FilterByOfferTypesContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 
 import FilterByOfferTypes from './FilterByOfferTypes'
-import selectTypeSublabels from '../selectors/selectTypes'
+import selectTypeSublabels from '../../../../selectors/data/typesSelector'
 
 const mapStateToProps = state => ({
   typeSublabels: selectTypeSublabels(state),
