+++ src/helpers/removePageFromSearchString.js
@@ -1,0 +1,7 @@
+const removePageFromSearchString = searchString =>
+  searchString
+    .split('&')
+    .filter(element => !element.includes('page='))
+    .join('&')
+
+export default removePageFromSearchString
