+++ src/selectors/data/__specs__/stocksSelector.spec.js
@@ -1,275 +1,273 @@
 import { selectIsEnoughStockForOfferDuo, selectIsStockDuo } from '../stocksSelector'
 
-describe('src | components | layout | Duo | selectors | stocksSelector', () => {
-  describe('selectIsEnoughStockForOfferDuo', () => {
-    it('should return true when there is stock with more than two available', () => {
-      // given
-      const state = {
-        data: {
-          stocks: [
-            {
-              offerId: 'AAAA',
-              available: 2,
-            },
-          ],
-        },
-      }
+describe('selectIsEnoughStockForOfferDuo', () => {
+  it('should return true when there is stock with more than two available', () => {
+    // given
+    const state = {
+      data: {
+        stocks: [
+          {
+            offerId: 'AAAA',
+            available: 2,
+          },
+        ],
+      },
+    }
 
-      const offerId = 'AAAA'
+    const offerId = 'AAAA'
 
-      // when
-      const isEnoughStockForOfferDuo = selectIsEnoughStockForOfferDuo(state, offerId)
+    // when
+    const isEnoughStockForOfferDuo = selectIsEnoughStockForOfferDuo(state, offerId)
 
-      // then
-      expect(isEnoughStockForOfferDuo).toBe(true)
-    })
-
-    it('should return true when there is unlimited stock', () => {
-      // given
-      const state = {
-        data: {
-          stocks: [
-            {
-              offerId: 'AAAA',
-              available: null,
-            },
-          ],
-        },
-      }
-
-      const offerId = 'AAAA'
-
-      // when
-      const isEnoughStockForOfferDuo = selectIsEnoughStockForOfferDuo(state, offerId)
-
-      // then
-      expect(isEnoughStockForOfferDuo).toBe(true)
-    })
-
-    it('should return false when there is no stock with more than two available', () => {
-      // given
-      const state = {
-        data: {
-          stocks: [
-            {
-              offerId: 'AAAA',
-              available: 1,
-            },
-          ],
-        },
-      }
-
-      const offerId = 'AAAA'
-
-      // when
-      const isEnoughStockForOfferDuo = selectIsEnoughStockForOfferDuo(state, offerId)
-
-      // then
-      expect(isEnoughStockForOfferDuo).toBe(false)
-    })
-
-    it('should return false when there is no match between stock and offer', () => {
-      // given
-      const state = {
-        data: {
-          stocks: [
-            {
-              offerId: 'AAAA',
-              available: 5,
-            },
-          ],
-        },
-      }
-
-      const offerId = 'BBBB'
-
-      // when
-      const isEnoughStockForOfferDuo = selectIsEnoughStockForOfferDuo(state, offerId)
-
-      // then
-      expect(isEnoughStockForOfferDuo).toBe(false)
-    })
+    // then
+    expect(isEnoughStockForOfferDuo).toBe(true)
   })
 
-  describe('selectIsStockDuo', () => {
-    it('should be duo if feature is active, stock available 2 or more and on an offer duo', () => {
-      // given
-      const state = {
-        data: {
-          features: [
-            {
-              nameKey: 'DUO_OFFER',
-              isActive: true,
-            },
-          ],
-          offers: [
-            {
-              id: 'O1',
-              isDuo: true,
-            },
-          ],
-          stocks: [
-            {
-              id: 'S1',
-              offerId: 'O1',
-              available: 5,
-            },
-          ],
-        },
-      }
+  it('should return true when there is unlimited stock', () => {
+    // given
+    const state = {
+      data: {
+        stocks: [
+          {
+            offerId: 'AAAA',
+            available: null,
+          },
+        ],
+      },
+    }
 
-      const offerId = 'O1'
-      const stockId = 'S1'
+    const offerId = 'AAAA'
 
-      // when
-      const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
+    // when
+    const isEnoughStockForOfferDuo = selectIsEnoughStockForOfferDuo(state, offerId)
 
-      // then
-      expect(isStockDuo).toBe(true)
-    })
+    // then
+    expect(isEnoughStockForOfferDuo).toBe(true)
+  })
 
-    it('should be duo if feature is active, stock unlimited and on an offer duo', () => {
-      // given
-      const state = {
-        data: {
-          features: [
-            {
-              nameKey: 'DUO_OFFER',
-              isActive: true,
-            },
-          ],
-          offers: [
-            {
-              id: 'O1',
-              isDuo: true,
-            },
-          ],
-          stocks: [
-            {
-              id: 'S1',
-              offerId: 'O1',
-              available: null,
-            },
-          ],
-        },
-      }
+  it('should return false when there is no stock with more than two available', () => {
+    // given
+    const state = {
+      data: {
+        stocks: [
+          {
+            offerId: 'AAAA',
+            available: 1,
+          },
+        ],
+      },
+    }
 
-      const offerId = 'O1'
-      const stockId = 'S1'
+    const offerId = 'AAAA'
 
-      // when
-      const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
+    // when
+    const isEnoughStockForOfferDuo = selectIsEnoughStockForOfferDuo(state, offerId)
 
-      // then
-      expect(isStockDuo).toBe(true)
-    })
+    // then
+    expect(isEnoughStockForOfferDuo).toBe(false)
+  })
 
-    it('should not be duo if feature is not active', () => {
-      // given
-      const state = {
-        data: {
-          features: [
-            {
-              nameKey: 'DUO_OFFER',
-              isActive: false,
-            },
-          ],
-          offers: [
-            {
-              id: 'O1',
-              isDuo: true,
-            },
-          ],
-          stocks: [
-            {
-              id: 'S1',
-              offerId: 'O1',
-              available: 5,
-            },
-          ],
-        },
-      }
+  it('should return false when there is no match between stock and offer', () => {
+    // given
+    const state = {
+      data: {
+        stocks: [
+          {
+            offerId: 'AAAA',
+            available: 5,
+          },
+        ],
+      },
+    }
 
-      const offerId = 'O1'
-      const stockId = 'S1'
+    const offerId = 'BBBB'
 
-      // when
-      const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
+    // when
+    const isEnoughStockForOfferDuo = selectIsEnoughStockForOfferDuo(state, offerId)
 
-      // then
-      expect(isStockDuo).toBe(false)
-    })
+    // then
+    expect(isEnoughStockForOfferDuo).toBe(false)
+  })
+})
 
-    it('should not be duo if there less than 2 available', () => {
-      // given
-      const state = {
-        data: {
-          features: [
-            {
-              nameKey: 'DUO_OFFER',
-              isActive: true,
-            },
-          ],
-          offers: [
-            {
-              id: 'O1',
-              isDuo: true,
-            },
-          ],
-          stocks: [
-            {
-              id: 'S1',
-              offerId: 'O1',
-              available: 1,
-            },
-          ],
-        },
-      }
+describe('selectIsStockDuo', () => {
+  it('should be duo if feature is active, stock available 2 or more and on an offer duo', () => {
+    // given
+    const state = {
+      data: {
+        features: [
+          {
+            nameKey: 'DUO_OFFER',
+            isActive: true,
+          },
+        ],
+        offers: [
+          {
+            id: 'O1',
+            isDuo: true,
+          },
+        ],
+        stocks: [
+          {
+            id: 'S1',
+            offerId: 'O1',
+            available: 5,
+          },
+        ],
+      },
+    }
 
-      const offerId = 'O1'
-      const stockId = 'S1'
+    const offerId = 'O1'
+    const stockId = 'S1'
 
-      // when
-      const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
+    // when
+    const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
 
-      // then
-      expect(isStockDuo).toBe(false)
-    })
+    // then
+    expect(isStockDuo).toBe(true)
+  })
 
-    it('should not be duo if his offer is not duo', () => {
-      // given
-      const state = {
-        data: {
-          features: [
-            {
-              nameKey: 'DUO_OFFER',
-              isActive: true,
-            },
-          ],
-          offers: [
-            {
-              id: 'O1',
-              isDuo: false,
-            },
-          ],
-          stocks: [
-            {
-              id: 'S1',
-              offerId: 'O1',
-              available: 2,
-            },
-          ],
-        },
-      }
+  it('should be duo if feature is active, stock unlimited and on an offer duo', () => {
+    // given
+    const state = {
+      data: {
+        features: [
+          {
+            nameKey: 'DUO_OFFER',
+            isActive: true,
+          },
+        ],
+        offers: [
+          {
+            id: 'O1',
+            isDuo: true,
+          },
+        ],
+        stocks: [
+          {
+            id: 'S1',
+            offerId: 'O1',
+            available: null,
+          },
+        ],
+      },
+    }
 
-      const offerId = 'O1'
-      const stockId = 'S1'
+    const offerId = 'O1'
+    const stockId = 'S1'
 
-      // when
-      const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
+    // when
+    const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
 
-      // then
-      expect(isStockDuo).toBe(false)
-    })
+    // then
+    expect(isStockDuo).toBe(true)
+  })
+
+  it('should not be duo if feature is not active', () => {
+    // given
+    const state = {
+      data: {
+        features: [
+          {
+            nameKey: 'DUO_OFFER',
+            isActive: false,
+          },
+        ],
+        offers: [
+          {
+            id: 'O1',
+            isDuo: true,
+          },
+        ],
+        stocks: [
+          {
+            id: 'S1',
+            offerId: 'O1',
+            available: 5,
+          },
+        ],
+      },
+    }
+
+    const offerId = 'O1'
+    const stockId = 'S1'
+
+    // when
+    const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
+
+    // then
+    expect(isStockDuo).toBe(false)
+  })
+
+  it('should not be duo if there less than 2 available', () => {
+    // given
+    const state = {
+      data: {
+        features: [
+          {
+            nameKey: 'DUO_OFFER',
+            isActive: true,
+          },
+        ],
+        offers: [
+          {
+            id: 'O1',
+            isDuo: true,
+          },
+        ],
+        stocks: [
+          {
+            id: 'S1',
+            offerId: 'O1',
+            available: 1,
+          },
+        ],
+      },
+    }
+
+    const offerId = 'O1'
+    const stockId = 'S1'
+
+    // when
+    const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
+
+    // then
+    expect(isStockDuo).toBe(false)
+  })
+
+  it('should not be duo if his offer is not duo', () => {
+    // given
+    const state = {
+      data: {
+        features: [
+          {
+            nameKey: 'DUO_OFFER',
+            isActive: true,
+          },
+        ],
+        offers: [
+          {
+            id: 'O1',
+            isDuo: false,
+          },
+        ],
+        stocks: [
+          {
+            id: 'S1',
+            offerId: 'O1',
+            available: 2,
+          },
+        ],
+      },
+    }
+
+    const offerId = 'O1'
+    const stockId = 'S1'
+
+    // when
+    const isStockDuo = selectIsStockDuo(state, stockId, offerId, 'DUO_OFFER')
+
+    // then
+    expect(isStockDuo).toBe(false)
   })
 })
